+++ app/code/community/Ecocode/Profiler/Model/LoggerInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+interface Ecocode_Profiler_Model_LoggerInterface
+ extends Ecocode_Profiler_Model_Logger_DebugHandlerInterface
+{
+    public function mageLog($level, $message, array $context = []);
+}